Split boardinit.ts off gamefile.ts
